# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSBUTools API",
        "description": "RESTful API providing data related to Super Smash Bros. Ultimate assets",
        "contact": {
            "name": "Raynald Mirville",
            "email": "admin@ssbutools.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/v1",
            "description": "SSBUTools API Server"
        }
    ],
    "paths": {
        "/stages/classifications": {
            "get": {
                "summary": "Returns an index of stage classifications",
                "description": "Returns all of stage classifications",
                "operationId": "classificationsIndex",
                "responses": {
                    "200": {
                        "description": "An index of all stage classifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_links": {
                                            "$ref": "#/components/schemas/links_index"
                                        },
                                        "_embedded": {
                                            "properties": {
                                                "classifications": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/stage_summary"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages/{id}/classifications": {
            "get": {
                "summary": "Returns a stage's classifications",
                "description": "Returns a stage's classifications",
                "operationId": "classificationsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of stage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A stage's classifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stage_classification"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages": {
            "get": {
                "summary": "Returns an index of stages",
                "description": "Returns all stages",
                "operationId": "stagesIndex",
                "responses": {
                    "200": {
                        "description": "An index of all stages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_links": {
                                            "properties": {
                                                "self": {
                                                    "$ref": "#/components/schemas/link"
                                                },
                                                "classifications": {
                                                    "$ref": "#/components/schemas/link"
                                                },
                                                "classificationSets": {
                                                    "$ref": "#/components/schemas/link"
                                                },
                                                "gameData": {
                                                    "$ref": "#/components/schemas/link"
                                                },
                                                "pieceMaps": {
                                                    "$ref": "#/components/schemas/link"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "_embedded": {
                                            "properties": {
                                                "stages": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/stage_summary"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages/{id}": {
            "get": {
                "summary": "Returns a stage's index of operations",
                "description": "Returns a stage's index of operations",
                "operationId": "stagesShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of stage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A stage's index of operations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stage"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages/game-data": {
            "get": {
                "summary": "Returns an index of stages' game data",
                "description": "Returns a summary of all stages' game data",
                "operationId": "stageGameDataIndex",
                "responses": {
                    "200": {
                        "description": "An index of all stages' game data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_links": {
                                            "$ref": "#/components/schemas/links_index"
                                        },
                                        "_embedded": {
                                            "properties": {
                                                "gameData": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "_links": {
                                                                "$ref": "#/components/schemas/links"
                                                            },
                                                            "id": {
                                                                "$ref": "#/components/schemas/id"
                                                            },
                                                            "name": {
                                                                "$ref": "#/components/schemas/name"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages/{id}/game-data": {
            "get": {
                "summary": "Returns a stage's game data",
                "description": "Returns a stage's game data",
                "operationId": "stageGameDataShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of stage",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A stage's game data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stage_game_data"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "link": {
                "properties": {
                    "href": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "links": {
                "properties": {
                    "self": {
                        "$ref": "#/components/schemas/link"
                    }
                },
                "type": "object"
            },
            "links_index": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/links"
                    },
                    {
                        "properties": {
                            "index": {
                                "$ref": "#/components/schemas/link"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "links_stage": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/links_index"
                    },
                    {
                        "properties": {
                            "stage": {
                                "$ref": "#/components/schemas/link"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "error": {
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "error_response": {
                "properties": {
                    "_links": {
                        "properties": {
                            "home": {
                                "$ref": "#/components/schemas/link"
                            }
                        },
                        "type": "object"
                    },
                    "_error": {
                        "$ref": "#/components/schemas/error"
                    }
                },
                "type": "object"
            },
            "id": {
                "type": "string"
            },
            "game_name": {
                "description": "The stage's in-game ID",
                "type": "string"
            },
            "name": {
                "description": "The stage's name",
                "type": "string"
            },
            "stage_base": {
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "gameName": {
                        "$ref": "#/components/schemas/game_name"
                    },
                    "name": {
                        "$ref": "#/components/schemas/name"
                    }
                },
                "type": "object"
            },
            "stage": {
                "allOf": [
                    {
                        "properties": {
                            "_links": {
                                "properties": {
                                    "self": {
                                        "$ref": "#/components/schemas/link"
                                    },
                                    "index": {
                                        "$ref": "#/components/schemas/link"
                                    },
                                    "classifications": {
                                        "$ref": "#/components/schemas/link"
                                    },
                                    "gameData": {
                                        "$ref": "#/components/schemas/link"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/stage_base"
                    }
                ]
            },
            "stage_summary": {
                "allOf": [
                    {
                        "properties": {
                            "_links": {
                                "$ref": "#/components/schemas/links"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/stage_base"
                    }
                ]
            },
            "stage_classification": {
                "description": "The collection of descriptors of a stage",
                "properties": {
                    "_links": {
                        "$ref": "#/components/schemas/links_stage"
                    },
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "gameName": {
                        "$ref": "#/components/schemas/game_name"
                    },
                    "name": {
                        "$ref": "#/components/schemas/name"
                    },
                    "abbr": {
                        "description": "The stage's abbreviated name",
                        "type": "string"
                    },
                    "series": {
                        "description": "The stage's video game series of origin",
                        "type": "string"
                    },
                    "tourneyPresence": {
                        "description": "The stage's presence in major tournaments",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "stage_material": {
                "properties": {
                    "leftLedge": {
                        "type": "boolean"
                    },
                    "rightLedge": {
                        "type": "boolean"
                    },
                    "noWallJump": {
                        "type": "boolean"
                    },
                    "passthroughAngle": {
                        "type": "number"
                    },
                    "length": {
                        "type": "number"
                    },
                    "ceiling": {
                        "type": "boolean"
                    },
                    "wall": {
                        "type": "boolean"
                    },
                    "material": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "stage_piece": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "vertex": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    },
                    "materials": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/stage_material"
                        }
                    },
                    "boundingBox": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "stage_game_data_raw": {
                "properties": {
                    "stage": {
                        "type": "string"
                    },
                    "name": {
                        "$ref": "#/components/schemas/name"
                    },
                    "lvd": {
                        "type": "string"
                    },
                    "collisions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/stage_piece"
                        }
                    }
                },
                "type": "object"
            },
            "stage_game_data": {
                "properties": {
                    "_links": {
                        "$ref": "#/components/schemas/links_stage"
                    },
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/stage_game_data_raw"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "resource_not_found": {
                "description": "Requested resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error_response"
                        }
                    }
                }
            },
            "server_error": {
                "description": "Unexpected error occured",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error_response"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}